Improve testing code readability
